+++ lib/src/create_error.dart
@@ -9,6 +9,11 @@ class CreateError implements HttpException {
   /// HTTP status code (defaults to 400).
   final int errorCode;
 
+  ///  * Creates a new `Error` that can be used to handle both internal and runtime errors.
+  ///
+  ///  * Pass an error message
+  ///
+  /// * Pass a http status code (defaults to 400) (optional)
   CreateError({
     required this.message,
     this.errorCode = 400,